+++ src/index.ts
@@ -1,11 +1,11 @@
-export { AckModule } from "./AckModule";
-export { RouteWatchReporter } from "./RouteWatchReporter";
+export { AckModule } from "./AckModule"
+export { RouteWatchReporter } from "./RouteWatchReporter"
 
-export { declarations as components } from "./declarations";
-export { declarations as pipes } from "./pipes";
+export { declarations as components } from "./declarations"
+export { declarations as pipes } from "./pipes"
 
 export { ack } from "ack-x/browser"
+export { AckApi } from "./providers/AckApi"
+export { AckArray } from "./directives/AckArray.directive"
 
-export { AckArray } from "./directives/AckArray.directive";
-
-export * from "./providers";
+export * from "./providers"